Rename BarInfo method to match method's naming requirements
