+++ social-tap/social-tap.csproj
@@ -59,6 +59,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BarInfo.cs" />
     <Compile Include="Comparer.cs" />
     <Compile Include="PixelCounter.cs" />
     <Compile Include="Form1.cs">
